robotics/analog_sensor: use strong typing
Turn 'int' type into 'uint16_t' type as it is always unsigned values.
As sum variable can overflow uint16_t before the average is done, it is
declared as uint32_t.

+++ robotics/include/analog_sensor.h
@@ -6,7 +6,7 @@
 #include <limits.h>
 
 #define AS_DIST_MAX	UINT8_MAX
-typedef int dist_cm_t;
+typedef uint8_t dist_cm_t;
 
 #define AS_DIST_LIMIT	50 /*cm*/
 
@@ -29,7 +29,7 @@ typedef struct {
 	double angle_robot_offset;
 
 	/* acquisition context */
-	int raw_values[ANALOG_SENSOR_NB_SAMPLES]; /* keep acquired distances */
+	uint16_t raw_values[ANALOG_SENSOR_NB_SAMPLES]; /* keep acquired distances */
 } sensor_t;
 
 typedef struct {
